feat: Add new home page sections, core UI components, and the crowd-check page.
